rename as chimeSelectors
